expansao do sistema de cadastramento de pacientes, medico, consulta, diagnostico, internacao. outras coisas tmb etc tudo e tal

+++ src/Leitura.java
@@ -12,7 +12,7 @@ public class Leitura {
 		return this.ctt_int;
 	}
 	
-	public int getFlo() {
+	public float getFlo() {
 		return this.ctt_flo;
 	}
